- fix missing include in recint - fix cast bug in init of modular<int64>

+++ src/kernel/ring/modular-int64.inl
@@ -385,8 +385,8 @@ namespace Givaro {
     inline  typename Modular<int64_t, COMP>::Element&  Modular<int64_t, COMP>::init ( Element& r, const long long a ) const
 	{
 		int sign; uint64_t ua;
-		if (a <0) { sign =-1; ua = (unsigned int)-a;}
-		else { ua = (unsigned int)a; sign =1; }
+		if (a <0) { sign =-1; ua = (uint64_t)-a;}
+		else { ua = (uint64_t)a; sign =1; }
 		r = (Element) ( (ua >=_p) ? ua % (uint64_t)_p : ua) ;
 		if (r && (sign ==-1)) r = (Element)_p - r;
                 return r;
